Start with empirical ocean simulation
