+++ src/math/mod.rs
@@ -4,29 +4,15 @@ use na;
 use ndarray::{Array, ArrayBase, ArrayView, ArrayViewMut, Dimension, Ix1};
 use generic_array::ArrayLength;
 
+pub mod interp;
 pub mod vector_n;
 pub mod wavelet;
 
+pub use self::interp::{linear, bilinear, trilinear};
 pub use self::vector_n::VectorN;
 
 pub fn vec2<N: na::Scalar>(x: N, y: N) -> na::Vector2<N> {
     na::Vector2::new(x, y)
-}
-
-pub fn linear<S: Real>(a0: S, a1: S, s: S) -> S {
-    a0 * (S::one() - s) + a1 * s
-}
-
-pub fn bilinear<S: Real>(a00: S, a01: S, a10: S, a11: S, s: S, t: S) -> S {
-    linear(linear(a00, a01, s), linear(a10, a11, s), t)
-}
-
-pub fn trilinear<S: Real>(
-    a000: S, a001: S, a010: S, a011: S,
-    a100: S, a101: S, a110: S, a111: S,
-    s: S, t: S, u: S
-) -> S {
-    linear(bilinear(a000, a001, a010, a011, s, t), bilinear(a100, a101, a110, a111, s, t), u)
 }
 
 pub trait LinearView {
