+++ src/ocean/mod.rs
@@ -1,0 +1,11 @@
+
+//! Realistic ocean simulation
+//!
+//! Reference:
+//!     [Hoarth15] Christopher J. Horvath, 2015,
+//!                Empirical directional wave spectra for computer graphics,
+//!                In Proceedings of the 2015 Symposium on Digital Production (DigiPro '15),
+//!                Stephen Spencer (Ed.). ACM, New York, NY, USA, 29-39,
+//!                DOI=http://dx.doi.org/10.1145/2791261.2791267
+
+pub mod empirical;
